Added method loading depending on the simulation
